Move functions to helper headers
