+++ interfaces/Make.package
@@ -19,7 +19,9 @@ CEXE_sources += eos_data.cpp
 endif
 
 CEXE_headers += network.H
+CEXE_headers += network_utilities.H
 ifeq ($(USE_CXX_REACTIONS),TRUE)
+CEXE_headers += rhs_utilities.H
 CEXE_sources += network_initialization.cpp
 CEXE_headers += fortran_to_cxx_actual_rhs.H
 F90EXE_sources += fortran_to_cxx_actual_rhs.F90
